bugfixes
вынесены в отельный файл, функции для преобразования входящих данных (жанры, дата),
в поиске добавлены жанры и даты,
в поиске временно закомментировано добавление найденных фильмов в localStorage
