+++ src/js/markupHome.js
@@ -2,6 +2,9 @@
 import refs from './refs/refs';
 import { getFilm } from './modalFilm.js';
 import movieTemplate from '../templates/film-card-template.hbs';
+import { toGetShortGenresList } from './data-converting-functions.js';
+import { toGetFullGenresList } from './data-converting-functions.js';
+import { toGetYear } from './data-converting-functions.js';
 
 const trendingApiServise = window.ApiService;
 
@@ -45,59 +48,59 @@ function toSaveGenres(data) {
   localStorage.setItem('genres', JSON.stringify(data));
 }
 
-// transforming full date in year in results
-function toGetYear(data) {
-  data.results.map(item => {
-    item.release_year = item.release_date.slice(0, 4);
-  });
+// // transforming full date in year in results
+// export function toGetYear(data) {
+//   data.results.map(item => {
+//     item.release_year = item.release_date.slice(0, 4);
+//   });
 
-  data.results.map(item => {
-    if (!item.release_date) {
-      item.release_year = 'No release year';
-    }
-  });
+//   data.results.map(item => {
+//     if (!item.release_date) {
+//       item.release_year = 'No release year';
+//     }
+//   });
 
-  return data;
-}
+//   return data;
+// }
 
-// getting short genres names list from local storage
-function toGetShortGenresList(data) {
-  const genres = JSON.parse(localStorage.getItem('genres'));
-  data.results.map(item => {
-    let filmGenres = [];
-    genres.find(elem => {
-      if (item.genre_ids.includes(elem.id)) {
-        filmGenres.push(elem.name);
-      }
-    });
+// // getting short genres names list from local storage
+// export function toGetShortGenresList(data) {
+//   const genres = JSON.parse(localStorage.getItem('genres'));
+//   data.results.map(item => {
+//     let filmGenres = [];
+//     genres.find(elem => {
+//       if (item.genre_ids.includes(elem.id)) {
+//         filmGenres.push(elem.name);
+//       }
+//     });
 
-    if (filmGenres.length <= 3) {
-      item.genresShort = filmGenres.join(', ');
-    }
-    if (filmGenres.length > 3) {
-      filmGenres.splice(2, filmGenres.length - 2);
-      filmGenres.push('Other');
-      item.genresShort = filmGenres.join(', ');
-    }
-    if (filmGenres.length === 0) {
-      filmGenres.push('No genres');
-      item.genresShort = filmGenres.join(', ');
-    }
-  });
-  return data;
-}
+//     if (filmGenres.length <= 3) {
+//       item.genresShort = filmGenres.join(', ');
+//     }
+//     if (filmGenres.length > 3) {
+//       filmGenres.splice(2, filmGenres.length - 2);
+//       filmGenres.push('Other');
+//       item.genresShort = filmGenres.join(', ');
+//     }
+//     if (filmGenres.length === 0) {
+//       filmGenres.push('Other');
+//       item.genresShort = filmGenres.join(', ');
+//     }
+//   });
+//   return data;
+// }
 
-// getting full genres names list from local storage
-function toGetFullGenresList(data) {
-  const genres = JSON.parse(localStorage.getItem('genres'));
-  data.results.map(item => {
-    let filmGenresAll = [];
-    genres.find(elem => {
-      if (item.genre_ids.includes(elem.id)) {
-        filmGenresAll.push(elem.name);
-      }
-    });
-    item.genresAll = filmGenresAll;
-  });
-  return data;
-}
+// // getting full genres names list from local storage
+// export function toGetFullGenresList(data) {
+//   const genres = JSON.parse(localStorage.getItem('genres'));
+//   data.results.map(item => {
+//     let filmGenresAll = [];
+//     genres.find(elem => {
+//       if (item.genre_ids.includes(elem.id)) {
+//         filmGenresAll.push(elem.name);
+//       }
+//     });
+//     item.genresAll = filmGenresAll;
+//   });
+//   return data;
+// }
