+++ src/js/searchMovie.js
@@ -5,6 +5,9 @@ import movieTemplate from '../templates/film-card-template.hbs';
 import { myNotice, myError, myAlert } from './components/pnotify';
 import refs from './refs/refs';
 import { getFilm } from './modalFilm.js';
+import { toGetShortGenresList } from './data-converting-functions.js';
+import { toGetFullGenresList } from './data-converting-functions.js';
+import { toGetYear } from './data-converting-functions.js';
 
 const nameOfMovieToSearch = window.ApiService;
 
@@ -61,6 +64,9 @@ function onSearch(event) {
 function fetchSearch() {
   nameOfMovieToSearch
     .fetchSearchMovies()
+    .then(toGetYear)
+    .then(toGetShortGenresList)
+    .then(toGetFullGenresList)
     .then(data => {
       getFilm(data.results);
       if (data.results.length > 0) {
@@ -85,12 +91,12 @@ function fetchSearch() {
     .catch(error => console.log(error));
 
   //==adding search result to the localStorage==
-  nameOfMovieToSearch
-    .fetchSearchMovies()
-    .then(result => {
-      localStorage.setItem('searchResult', JSON.stringify(result));
-    })
-    .catch(error => console.log(error));
+  // nameOfMovieToSearch
+  //   .fetchSearchMovies()
+  //   .then(result => {
+  //     localStorage.setItem('searchResult', JSON.stringify(result));
+  //   })
+  //   .catch(error => console.log(error));
 }
 
 function renderMakrup(results) {
@@ -101,5 +107,5 @@ function clearContent() {
   //==очистка содержимого страницы перед выведением результатов поиска===
   refs.gallery.innerHTML = '';
   nameOfMovieToSearch.resetPage;
-  localStorage.clear('searchResult');
+  // localStorage.clear('searchResult');
 }
